Integrate new checkpoint (#185)
* integrate new checkpoint system

* format code

* add error logging

* add more logging

* add epoch of fund and release

* add more logging

* update info

* update bottom up

* update comment

* update log

* update logging

* remove subsystem file
