+++ src/manager/checkpoint/mod.rs
@@ -1,38 +1,33 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: MIT
 
-//! Issues:
-//! 1. Checkpoint periods are fetched
-use std::collections::HashMap;
-use std::sync::Arc;
-use std::time::Duration;
-
-use anyhow::Result;
+use crate::config::{ReloadableConfig, Subnet};
+use crate::lotus::client::{DefaultLotusJsonRPCClient, LotusJsonRPCClient};
+use crate::lotus::LotusClient;
+use crate::manager::checkpoint::manager::bottomup::BottomUpCheckpointManager;
+use crate::manager::checkpoint::manager::topdown::TopDownCheckpointManager;
+use crate::manager::checkpoint::manager::CheckpointManager;
+use anyhow::{anyhow, Result};
 use async_trait::async_trait;
-use futures_util::stream::FuturesUnordered;
-use futures_util::StreamExt;
+use cid::Cid;
+use futures_util::future::join_all;
+use fvm_shared::address::Address;
+use fvm_shared::clock::ChainEpoch;
 use ipc_sdk::subnet_id::SubnetID;
+use std::collections::HashMap;
+use std::str::FromStr;
+use std::sync::Arc;
+use std::time::{Duration, Instant};
 use tokio::select;
-use tokio::sync::Notify;
 use tokio::time::sleep;
 use tokio_graceful_shutdown::{IntoSubsystem, SubsystemHandle};
 
-use crate::config::{ReloadableConfig, Subnet};
-use bottomup::manage_bottomup_checkpoints;
-use topdown::manage_topdown_checkpoints;
-
 pub use manager::*;
-pub use subsystem::*;
-
-pub(crate) mod bottomup;
 mod manager;
-mod subsystem;
-pub(crate) mod topdown;
 
-/// The frequency at which to check a new chain head.
-pub(crate) const CHAIN_HEAD_REQUEST_PERIOD: Duration = Duration::from_secs(10);
+const TASKS_PROCESS_THRESHOLD_SEC: u64 = 15;
+const SUBMISSION_LOOK_AHEAD_EPOCH: ChainEpoch = 50;
 
-/// The `CheckpointSubsystem`. When run, it actively monitors subnets and submits checkpoints.
 pub struct CheckpointSubsystem {
     /// The subsystem uses a `ReloadableConfig` to ensure that, at all, times, the subnets under
     /// management are those in the latest version of the config.
@@ -48,106 +43,227 @@ impl CheckpointSubsystem {
 
 #[async_trait]
 impl IntoSubsystem<anyhow::Error> for CheckpointSubsystem {
-    /// Runs the checkpoint subsystem, which actively monitors subnets and submits checkpoints.
-    /// For each (account, subnet) that exists in the config, the subnet is monitored and checkpoints
-    /// are submitted at the appropriate epochs.
-    async fn run(self, subsys: SubsystemHandle) -> Result<()> {
+    async fn run(self, subsys: SubsystemHandle) -> anyhow::Result<()> {
         // Each event in this channel is notification of a new config.
         let mut config_chan = self.config.new_subscriber();
 
         loop {
             // Load the latest config.
             let config = self.config.get_config();
+            let (top_down_managers, bottom_up_managers) =
+                setup_managers_from_config(&config.subnets).await?;
 
-            // Create a top-down and a bottom-up checkpoint manager future for each (child, parent)
-            // subnet pair under and collect them in a `FuturesUnordered` set.
-            let mut manage_subnet_futures = FuturesUnordered::new();
-            let stop_subnet_managers = Arc::new(Notify::new());
-            let subnets_to_manage = subnets_to_manage(&config.subnets);
-            log::debug!("We have {} subnets to manage", subnets_to_manage.len());
-
-            for (child, parent) in subnets_to_manage {
-                manage_subnet_futures.push(tokio::spawn(manage_bottomup_checkpoints(
-                    (child.clone(), parent.clone()),
-                    stop_subnet_managers.clone(),
-                )));
-                manage_subnet_futures.push(tokio::spawn(manage_topdown_checkpoints(
-                    (child.clone(), parent.clone()),
-                    stop_subnet_managers.clone(),
-                )));
-            }
-
-            // Spawn a task to drive the `manage_subnet` futures.
-            let manage_subnets_task = tokio::spawn(async move {
-                loop {
-                    match manage_subnet_futures.next().await {
-                        Some(Err(e)) => {
-                            panic!("Panic in manage_subnet: {:#}", e);
+            loop {
+                select! {
+                    _ = process_managers(&top_down_managers) => {},
+                    _ = process_managers(&bottom_up_managers) => {},
+                    r = config_chan.recv() => {
+                        log::info!("Config changed, reloading checkpointing subsystem");
+                        match r {
+                            // Config updated, return to caller
+                            Ok(_) => { break; },
+                            Err(_) => {
+                                return Err(anyhow!("Config channel unexpectedly closed, shutting down checkpointing subsystem"))
+                            },
                         }
-                        Some(Ok(r)) => {
-                            if let Err(e) = r {
-                                log::error!("Error in manage_subnet: {:#}", e);
-                            }
-                        }
-                        None => {
-                            log::debug!("All manage_subnet futures have finished");
-                            break;
-                        }
+                    }
+                    _ = subsys.on_shutdown_requested() => {
+                        log::info!("Shutting down checkpointing subsystem");
+                        return Ok(());
                     }
                 }
-            });
-
-            // Watch for shutdown requests and config changes.
-            let is_shutdown = select! {
-                _ = subsys.on_shutdown_requested() => {
-                    log::info!("Shutting down checkpointing subsystem");
-                    true
-                },
-                r = config_chan.recv() => {
-                    log::info!("Config changed, reloading checkpointing subsystem");
-                    match r {
-                        Ok(_) => { false },
-                        Err(_) => {
-                            log::error!("Config channel unexpectedly closed, shutting down checkpointing subsystem");
-                            true
-                        },
-                    }
-                },
-            };
-
-            // Cleanly stop the `manage_subnet` futures.
-            stop_subnet_managers.notify_waiters();
-            log::debug!("Waiting for subnet managers to finish");
-            manage_subnets_task.await?;
-
-            if is_shutdown {
-                return anyhow::Ok(());
             }
         }
     }
 }
 
-/// This function takes a `HashMap<String, Subnet>` and returns a `Vec` of tuples of the form
-/// `(child_subnet, parent_subnet)`, where `child_subnet` is a subnet that we need to actively
-/// manage checkpoint for. This means that for each `child_subnet` there exists at least one account
-/// for which we need to submit checkpoints on behalf of to `parent_subnet`, which must also be
-/// present in the map.
-fn subnets_to_manage(subnets_by_id: &HashMap<SubnetID, Subnet>) -> Vec<(Subnet, Subnet)> {
-    // We filter for subnets that have at least one account and for which the parent subnet
-    // is also in the map, and map into a Vec of (child_subnet, parent_subnet) tuples.
-    subnets_by_id
-        .values()
-        .filter(|s| !s.accounts.is_empty())
-        .filter(|s| s.id.parent().is_some() && subnets_by_id.contains_key(&s.id.parent().unwrap()))
-        .map(|s| (s.clone(), subnets_by_id[&s.id.parent().unwrap()].clone()))
-        .collect()
+fn handle_err_response(manager: &impl CheckpointManager, response: anyhow::Result<()>) {
+    if response.is_err() {
+        log::error!("manger {manager:} had error: {:}", response.unwrap_err());
+    }
 }
 
-/// Sleeps for some time if stop_notify is not fired. It returns true to flag that we should move to the
-/// next iteration of the loop, while false informs that the loop should return.
-pub async fn wait_next_iteration(stop_notify: &Arc<Notify>, timeout: Duration) -> Result<bool> {
-    select! {
-        _ = sleep(timeout) => {Ok(true)}
-        _ = stop_notify.notified() => {Ok(false)}
+async fn setup_managers_from_config(
+    subnets: &HashMap<SubnetID, Subnet>,
+) -> Result<(
+    Vec<TopDownCheckpointManager>,
+    Vec<BottomUpCheckpointManager<DefaultLotusJsonRPCClient>>,
+)> {
+    let mut bottom_up_managers = vec![];
+    let mut top_down_managers = vec![];
+
+    for s in subnets.values() {
+        log::info!("config checkpoint manager for subnet: {:}", s.id);
+
+        // We filter for subnets that have at least one account and for which the parent subnet
+        // is also in the configuration.
+        if s.accounts.is_empty() {
+            log::info!("no accounts in subnet: {:}, not managing checkpoints", s.id);
+            continue;
+        }
+
+        let parent = if let Some(p) = s.id.parent() && subnets.contains_key(&p) {
+            subnets.get(&p).unwrap()
+        } else {
+            log::info!("subnet has no parent configured: {:}, not managing checkpoints", s.id);
+            continue
+        };
+
+        bottom_up_managers.push(
+            BottomUpCheckpointManager::new(
+                LotusJsonRPCClient::from_subnet(parent),
+                parent.clone(),
+                LotusJsonRPCClient::from_subnet(s),
+                s.clone(),
+            )
+            .await?,
+        );
+
+        top_down_managers.push(
+            TopDownCheckpointManager::new(
+                LotusJsonRPCClient::from_subnet(parent),
+                parent.clone(),
+                LotusJsonRPCClient::from_subnet(s),
+                s.clone(),
+            )
+            .await?,
+        );
+    }
+
+    log::info!(
+        "we are managing checkpoints for {:} number of bottom up subnets",
+        bottom_up_managers.len()
+    );
+    log::info!(
+        "we are managing checkpoints for {:} number of top down subnets",
+        top_down_managers.len()
+    );
+
+    Ok((top_down_managers, bottom_up_managers))
+}
+
+async fn process_managers<T: CheckpointManager>(managers: &[T]) -> anyhow::Result<()> {
+    // Tracks the start time of the processing, will use this to determine should sleep
+    let start_time = Instant::now();
+
+    let futures = managers
+        .iter()
+        .map(|manager| async {
+            let response = submit_till_current_epoch(manager).await;
+            handle_err_response(manager, response);
+        })
+        .collect::<Vec<_>>();
+
+    join_all(futures).await;
+
+    sleep_or_continue(start_time).await;
+
+    Ok(())
+}
+
+async fn sleep_or_continue(start_time: Instant) {
+    let elapsed = start_time.elapsed().as_secs();
+    if elapsed < TASKS_PROCESS_THRESHOLD_SEC {
+        sleep(Duration::from_secs(TASKS_PROCESS_THRESHOLD_SEC - elapsed)).await
+    }
+}
+
+/// Attempts to submit checkpoints from the last executed epoch all the way to the current epoch for
+/// all the validators in the provided manager.
+async fn submit_till_current_epoch(manager: &impl CheckpointManager) -> Result<()> {
+    if !manager.presubmission_check().await? {
+        log::info!("subnet in manager: {manager:} not ready to submit checkpoint");
+        return Ok(());
+    }
+
+    // we might have to obtain the list of validators as some validators might leave the subnet
+    // we can improve the performance by caching if this slows down the process significantly.
+    let validators = obtain_validators(manager).await?;
+    let period = manager.checkpoint_period();
+
+    let last_executed_epoch = manager.last_executed_epoch().await?;
+    let current_epoch = manager.current_epoch().await?;
+
+    log::debug!(
+        "latest epoch {:?}, last executed epoch: {:?} for checkpointing: {:}",
+        current_epoch,
+        last_executed_epoch,
+        manager,
+    );
+
+    let mut next_epoch = last_executed_epoch + period;
+    let cut_off_epoch = std::cmp::min(
+        current_epoch,
+        SUBMISSION_LOOK_AHEAD_EPOCH + last_executed_epoch,
+    );
+
+    // Instead of loop all the way to `current_epoch`, we loop till `cut_off_epoch`.
+    // Reason because if the current epoch is significantly greater than last_executed_epoch and there
+    // are lots of validators in the network, loop all the way to current epoch might have some outdated
+    // data. Set a cut off epoch such that validators can sync with chain more regularly.
+    while next_epoch < cut_off_epoch {
+        // now we process each validator
+        for validator in &validators {
+            log::debug!("submit checkpoint for validator: {validator:?} in manager: {manager:}");
+
+            if !manager
+                .should_submit_in_epoch(validator, next_epoch)
+                .await?
+            {
+                log::debug!(
+                    "next submission epoch {next_epoch:?} already voted for validator: {:?} in manager: {manager:}",
+                    validator.to_string()
+                );
+                continue;
+            }
+
+            log::debug!(
+                "next submission epoch {next_epoch:} not voted for validator: {validator:} in manager: {manager:}, should vote"
+            );
+
+            manager.submit_checkpoint(next_epoch, validator).await?;
+
+            log::info!("checkpoint at epoch {next_epoch:} submitted for validator {validator:} in manager: {manager:}");
+        }
+
+        // increment next epoch
+        next_epoch += period;
+    }
+
+    log::info!("process checkpoint from epoch: {last_executed_epoch:} to {current_epoch:} in manager: {manager:}");
+
+    Ok(())
+}
+
+/// Obtain the validators in the subnet from the parent subnet of the manager
+async fn obtain_validators(manager: &impl CheckpointManager) -> anyhow::Result<Vec<Address>> {
+    let parent_client = manager.parent_client();
+    let parent_head = parent_client.chain_head().await?;
+
+    // A key assumption we make now is that each block has exactly one tip set. We panic
+    // if this is not the case as it violates our assumption.
+    // TODO: update this logic once the assumption changes (i.e., mainnet)
+    assert_eq!(parent_head.cids.len(), 1);
+
+    let cid_map = parent_head.cids.first().unwrap().clone();
+    let parent_tip_set = Cid::try_from(cid_map)?;
+    let child_subnet = manager.child_subnet();
+
+    let subnet_actor_state = parent_client
+        .ipc_read_subnet_actor_state(&child_subnet.id, parent_tip_set)
+        .await?;
+
+    match subnet_actor_state.validator_set.validators {
+        None => Ok(vec![]),
+        Some(validators) => {
+            let mut vs = vec![];
+            for v in validators {
+                let addr = Address::from_str(&v.addr)?;
+                if child_subnet.accounts.contains(&addr) {
+                    vs.push(addr);
+                }
+            }
+            Ok(vs)
+        }
     }
 }
