+++ src/manager/subnet.rs
@@ -47,23 +47,25 @@ pub trait SubnetManager {
         gateway_addr: Address,
     ) -> Result<HashMap<SubnetID, SubnetInfo>>;
 
-    /// Fund injects new funds from an account of the parent chain to a subnet
+    /// Fund injects new funds from an account of the parent chain to a subnet.
+    /// Returns the epoch that the fund is executed in the parent.
     async fn fund(
         &self,
         subnet: SubnetID,
         gateway_addr: Address,
         from: Address,
         amount: TokenAmount,
-    ) -> Result<()>;
+    ) -> Result<ChainEpoch>;
 
     /// Release creates a new check message to release funds in parent chain
+    /// Returns the epoch that the released is executed in the child.
     async fn release(
         &self,
         subnet: SubnetID,
         gateway_addr: Address,
         from: Address,
         amount: TokenAmount,
-    ) -> Result<()>;
+    ) -> Result<ChainEpoch>;
 
     /// Propagate a cross-net message forward
     async fn propagate(
